Add support for float type
This adds the float type (19) from Facebook Thrift. It is not part of
Apache Thrift, so it should not be used if interoperability is a concern.

Warning: this is a backwards incompatible change, as float was previously
supported with coercion to and from double. It is now a native type.